+ atom-animation component
